Move BMI metadata to separate file
